CASMCMS-7389: Transitioning to Github and Artifactory
As part of the transition to Github, CASM is pushing its artifacts to its
own artifactory. This commit references a base Docker image from this
new artifactory location, algol60. It also changes URLs to point to
Github rather than Stash.

Clone the cms-meta-tools directory to gain access to these scripts,
straight from the source. This eliminates installing them via RPM
into the build directory. It is cleaner and simpler.