CASMCMS-7753: Ignore CAPMC errors for nodes that fail to power off
CAPMC's behavior changed. It returns an error code when a node fails to
power off. Previously, it did not return an error. BOS handled this
situation by issuing a forceful power off request. This behavior change
causes BOS it ignore the failing node and not issue the forceful power
off request to it. The CAPMC change caused a BOS bug.

To combat this bug, BOS is going to ignore this error message. It will
not weed out the failing nodes and will attempt to issue them the
forceful power off request.