CASMCMS-8300:
In CASMCMS-5469: 'Keep rootfs parameters consistent' a change was made
to continue to meet the syntax expectations of the CPSS3 rootfs
provider. Specifically, to maintain the delimiters even when a rootfs
provide or rootfs provider ID (i.e. etag) were blank. This change was
made to the base RootFSProvider class. The change should have been made
to the CPSS3Provider class because it was specific to that class.

Because the change was made to the base class, it broke in Mercury
where they do not use the CPSS3Provider class.

This mod restores the original functionality and moves the change to the
CPSS3Provider class where it belongs.